magic. ya know
