+++ README.md
@@ -12,3 +12,7 @@ a #wowwwnetwork bot that detects and compliments selfies.
 🌊  a ~sea~ collaboration  🌊
 
 ✨ by [@swayandsea](http://twitter.com/swayandsea) ([bots](http://sui.sexy)) (compliments/branding) + [@colewillsea](http://twitter.com/colewillsea) ([bots](https://twitter.com/colewillsea/lists/my-robots/members)) (code) ✨
+
+-------------------------------------------------
+
+the popQueue function in [selfie_complimenter.js]() is modified/forked from [sorting-bot by Darius Kazemi](https://github.com/dariusk/sorting-bot/blob/master/index.js)
