+++ src/emoji.txt
