--- src/gentle_tweeter/README.md
@@ -1,9 +1,0 @@
-GENTLE_TWEETER
------------------------
-
-Gentle Tweeter pulls tweets off a redis queue and replies to them with a compliment,
-while also checking to make sure that it does not reply to the same user too often.
-
--------------------------------
-
-the popQueue is modified/forked from [sorting-bot by Darius Kazemi](https://github.com/dariusk/sorting-bot/blob/master/index.js)
