+++ src/selfie_complimenter.js
@@ -1,7 +1,7 @@
 var redis = require('redis')
 var client = redis.createClient()
 var compliment = require('./utils').compliment
-var config = require('../config')
+var config = require('./config')
 var twitConfig = config.twitter
 var replyInterval = config.replyInterval
 var Twit = require('twit')
