--- src/vigilant_listener/README.md
@@ -1,4 +1,0 @@
-VIGILANT_LISTENER
------------------------------
-
-Vigilant Listener streams the bot's timeline and publishes potentially-selfie-containing tweets to a redis channel.
