--- src/vigilant_listener/index.js
@@ -1,13 +1,0 @@
-var redis = require('redis')
-var client = redis.createClient()
-var thisTweetIsPromising = require('./utils').thisTweetIsPromising
-var Twit = require('twit')
-var config = require('../config').twitter
-var T = new Twit(config)
-
-var stream = T.stream('user')
-
-stream.on('tweet', function (t) {
-  console.log('processing', t.user.screen_name, t.text)
-  if (thisTweetIsPromising(t)) client.publish('OCVq', JSON.stringify(t))
-})
